requirements.txt
# Core API
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.9.1

# HTTP & async
httpx==0.28.1

# Configuration
python-dotenv==1.1.0
pyyaml==6.0.2

# Fuzzy string matching for vendor names (OCR-tolerant grading)
rapidfuzz==3.12.2

# OpenAI LLM agent (optional — requires OPENAI_API_KEY)
openai==1.82.0

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
anyio==4.9.0

# Optional: OCR from images (requires Tesseract binary installed separately)
# Pillow==11.2.1
# pytesseract==0.3.13
